float로 통일 일단은 더더더

+++ Sumdays/app/src/main/java/com/example/sumdays/data/repository/DailyEntryRepository.kt
@@ -47,7 +47,7 @@ class DailyEntryRepository (
         diary: String? = null,
         keywords: String? = null,
         aiComment: String? = null,
-        emotionScore: Double? = null,
+        emotionScore: Float? = null,
         emotionIcon: String? = null,
         themeIcon: String? = null
     ) {
